requirements.txt
anthropic>=0.40.0
langgraph>=0.2.50
langchain>=0.3.0
langchain-core>=0.3.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0